requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yfinance>=0.2.50
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
matplotlib>=3.8.0
python-dotenv>=1.0.1
groq>=0.18.0
